slight refactor of code structure
