+++ package.json
@@ -87,9 +87,9 @@
     "redis": "^4.6.4",
     "safe-stable-stringify": "^2.4.3",
     "snarkjs": "^0.6.1",
-    "warp-contracts": "^1.3.3",
+    "warp-contracts": "^1.4.1",
     "warp-contracts-lmdb": "^1.1.9",
-    "warp-contracts-plugin-deploy": "^1.0.0",
+    "warp-contracts-plugin-deploy": "^1.0.4",
     "warp-contracts-plugin-ethers": "^1.0.7",
     "warp-contracts-plugin-signature": "^1.0.10",
     "warp-contracts-plugin-snarkjs": "^0.1.3",
